requirements.txt
fastapi
pydantic
uvicorn
mysql-connector-python
requests
passlib[bcrypt]
python-jose
